# docs/chatgpt-action-openapi.yaml
openapi: 3.1.0
info:
  title: Be Blank Studio OS Agent API
  version: 1.0.0
  description: ChatGPT Action schema for creating Be Blank Studio OS projects in Firebase.
servers:
  - url: https://agent-euweanpera-as.a.run.app
    description: Deployed Be Blank Studio OS agent endpoint.
paths:
  /project:
    post:
      operationId: create_project
      summary: Create a Studio OS project
      description: |-
        Creates a Studio OS project from structured JSON or natural language extracted into JSON fields.

        Accepts natural language such as:
        - Karun Phuket deadline 15 June
        - สร้างโปรเจค Karun Phuket ลูกค้า Karun ที่ภูเก็ต ส่งงาน 15 มิ.ย.
      security:
        - AgentKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
            examples:
              structuredProject:
                summary: Structured create_project fields.
                value:
                  name: Karun Phuket
                  client: Karun
                  location: Phuket
                  handoverDate: "2026-06-15"
                  status: design
              naturalLanguageMapping:
                summary: Parse casual English text into project fields.
                description: 'User input: "Karun Phuket deadline 15 June"'
                value:
                  name: Karun Phuket
                  handoverDate: "2026-06-15"
              thaiNaturalLanguageMapping:
                summary: Parse casual Thai text into project fields.
                description: 'User input: "สร้างโปรเจค Karun Phuket ลูกค้า Karun ที่ภูเก็ต ส่งงาน 15 มิ.ย."'
                value:
                  name: Karun Phuket
                  client: Karun
                  location: Phuket
                  handoverDate: "2026-06-15"
                  status: design
      responses:
        "200":
          description: Project created or updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          description: Invalid command.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "401":
          description: Missing or invalid agent key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
components:
  securitySchemes:
    AgentKey:
      type: apiKey
      in: header
      name: x-agent-key
  schemas:
    CreateProjectRequest:
      type: object
      required:
        - name
      additionalProperties: true
      properties:
        name:
          type: string
          description: Project name.
        client:
          type: string
          description: Client name.
        location:
          type: string
          description: Project location.
        handoverDate:
          type: string
          format: date
          description: Target handover date.
        status:
          type: string
          description: Project status.
          enum:
            - concept
            - design
            - construction
            - handover
            - open
        notes:
          type: string
          description: Optional project notes.
        nextAction:
          type: string
          description: Optional next action.
    AgentResponse:
      type: object
      required:
        - ok
        - action
        - projectId
        - message
      properties:
        ok:
          type: boolean
        action:
          type: string
        projectId:
          type: string
        message:
          type: string